+++ config/config.exs
@@ -843,6 +843,19 @@ config :pleroma, Pleroma.Search.Elasticsearch.Cluster,
     }
   }
 
+config :pleroma, :translator,
+  enabled: false,
+  module: Akkoma.Translators.DeepL
+
+config :pleroma, :deepl,
+  # either :free or :pro
+  tier: :free,
+  api_key: ""
+
+config :pleroma, :libre_translate,
+  url: "http://127.0.0.1:5000",
+  api_key: nil
+
 # Import environment specific config. This must remain at the bottom
 # of this file so it overrides the configuration defined above.
 import_config "#{Mix.env()}.exs"