+++ lib/pleroma/akkoma/translators/translator.ex
@@ -1,0 +1,3 @@
+defmodule Pleroma.Akkoma.Translator do
+  @callback translate(String.t(), String.t()) :: {:ok, String.t(), String.t()} | {:error, any()}
+end